server/requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
debugpy==1.8.12
exceptiongroup==1.2.2
fastapi==0.115.8
fastapi-utils==0.8.0
h11==0.14.0
idna==3.10
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
mypy-extensions==1.0.0
psutil==5.9.8
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
typing-inspect==0.9.0
typing_extensions==4.12.2
uvicorn==0.34.0
APScheduler==3.11.0
selenium==4.29.0
requests==2.32.3
httpx==0.28.1
psycopg2-binary==2.9.10